+++ solutions/uri/1026/1026.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 int main() {
-    uint16_t a, b;
+    uint32_t a, b;
 
     while (std::cin >> a >> b) {
         std::cout << (a ^ b) << std::endl;
